+++ java/api/src/main/java/org/ray/api/Ray.java
@@ -2,6 +2,9 @@ package org.ray.api;
 
 import java.util.List;
 import org.ray.api.id.UniqueId;
+import org.ray.api.runtime.DefaultRayRuntimeFactory;
+import org.ray.api.runtime.RayRuntime;
+import org.ray.api.runtime.RayRuntimeFactory;
 
 /**
  * This class contains all public APIs of Ray.
