+++ java/api/src/main/java/org/ray/api/runtime/DefaultRayRuntimeFactory.java
@@ -1,4 +1,4 @@
-package org.ray.api;
+package org.ray.api.runtime;
 
 import java.lang.reflect.Method;
