+++ java/api/src/main/java/org/ray/api/runtime/RayRuntimeFactory.java
@@ -1,4 +1,4 @@
-package org.ray.api;
+package org.ray.api.runtime;
 
 /**
  * A factory that produces a RayRuntime instance.
